+++ src/generate-public-key.c
@@ -1,7 +1,7 @@
 #include "keys.h"
 
-PublicKey generatePublicKey(int p, int q, int e) {
-    int number = p * q;
+PublicKey generatePublicKey(long long p, long long q, long long e) {
+    long long number = p * q;
 
     return (PublicKey){ number, e };
 }
