+++ types/files.h
@@ -1,11 +1,11 @@
 #ifndef FILES_H
 #define FILES_H
 
-void createPublicKeyFile(int n, int e);
-void createCryptedDigitsFile(int *cryptedDigits, int size);
-void createDecryptedMessageFile(char *decryptedMessage, int size);
-void fillNumbersArray(int *numbersArray, char *file, int size);
-void fillDecryptedMessage(char *string, int size);
-int getAmountOfElements(char *file);
+void      createPublicKeyFile(long long n, long long e);
+void      createCryptedDigitsFile(long long* cryptedDigits, long long size);
+void      createDecryptedMessageFile(char* decryptedMessage, long long size);
+void      fillNumbersArray(long long* numbersArray, char* file, long long size);
+void      fillDecryptedMessage(char* string, long long size);
+long long getAmountOfElements(char* file);
 
 #endif
