+++ types/modular-exponentiation.h
@@ -1,6 +1,6 @@
 #ifndef MODULAR_EXPONENTIATION_H
 #define MODULAR_EXPONENTIATION_H
 
-int modularExponentiation(int base, int expo, int mod);
+long long modularExponentiation(long long base, long long expo, long long mod);
 
 #endif
