+++ types/modular-inverse.h
@@ -1,6 +1,6 @@
 #ifndef MODULAR_INVERSE_H
 #define MODULAR_INVERSE_H
 
-int modularInverse(int a, int m);
+long long modularInverse(long long a, long long m);
 
 #endif
